refactor: add Hero section component
