+++ src/pages/auth/Login.tsx
@@ -8,6 +8,7 @@ import {z} from 'zod'
 import loginSchema from '../../schema/login'
 import axios from 'axios'
 import { useAuth } from '../../hooks/useAuth'
+import BackButton from '../../components/BackButton'
 
 
 type formFields = z.infer<typeof loginSchema>
@@ -59,57 +60,60 @@ export default function Login() {
 
 
   return (
-    <div className="mx-4 md:flex md:flex-col md:items-center">
-      <div>
-        <img src={loginImage} alt="login-image" />
-      </div>
-      <div className="my-8 space-y-4">
-        <h1 className="text-secondary-blue text-2xl">
-          Let's <span className="font-bold">Sign In</span>
-        </h1>
-        <p className="text-gray-500">quis nostrud exercitation ullamco laboris nisi ut</p>
-      </div>
-      <form className="space-y-7 my-12 w-full max-w-sm" onSubmit={handleSubmit(onSubmit)}>
-        <div className="border border-black flex items-center justify-evenly p-2 rounded-lg">
-          <input
-            type="email"
-            placeholder="Email"
-            className="flex-1 outline-none"
-            {...register('email', { required: 'Email is required' })}
-          />
-          <MdOutgoingMail className="text-3xl" />
+    <div>
+      <BackButton />
+      <div className="mx-4 md:flex md:flex-col md:items-center">
+        <div>
+          <img src={loginImage} alt="login-image" />
         </div>
-        {errors && <div className="text-red-600">{errors.email?.message}</div>}
-        <div className="border border-black flex items-center justify-evenly p-2 rounded-lg">
-          <input
-            type="password"
-            placeholder="Password"
-            className="flex-1 outline-none"
-            {...register('password')}
-          />
-          <CiLock className="text-3xl" />
+        <div className="my-8 space-y-4">
+          <h1 className="text-secondary-blue text-2xl">
+            Let's <span className="font-bold">Sign In</span>
+          </h1>
+          <p className="text-gray-500">quis nostrud exercitation ullamco laboris nisi ut</p>
         </div>
-        {errors && <div className="text-red-600">{errors.password?.message}</div>}
-        <p className="cursor-pointer text-secondary-blue hover:text-blue-400">
-          <Link to="#">Forgot password?</Link>
+        <form className="space-y-7 my-12 w-full max-w-sm" onSubmit={handleSubmit(onSubmit)}>
+          <div className="border border-black flex items-center justify-evenly p-2 rounded-lg">
+            <input
+              type="email"
+              placeholder="Email"
+              className="flex-1 outline-none"
+              {...register('email', { required: 'Email is required' })}
+            />
+            <MdOutgoingMail className="text-3xl" />
+          </div>
+          {errors && <div className="text-red-600">{errors.email?.message}</div>}
+          <div className="border border-black flex items-center justify-evenly p-2 rounded-lg">
+            <input
+              type="password"
+              placeholder="Password"
+              className="flex-1 outline-none"
+              {...register('password')}
+            />
+            <CiLock className="text-3xl" />
+          </div>
+          {errors && <div className="text-red-600">{errors.password?.message}</div>}
+          <p className="cursor-pointer text-secondary-blue hover:text-blue-400">
+            <Link to="#">Forgot password?</Link>
+          </p>
+          {errors && <div className="text-red-600">{errors.root?.message}</div>}
+          <button
+            type="submit"
+            disabled={isSubmitting}
+            className="bg-green-primary-400 w-full rounded-lg py-3 text-white text-lg text-center cursor-pointer"
+          >
+            {isSubmitting ? 'Loading...' : 'Login'}
+          </button>
+        </form>
+        <p className="mt-10 mx-4">
+          Don’t have an account?
+          <Link to="/signup">
+            <span className="cursor-pointer text-secondary-blue mx-2 hover:text-blue-400">
+              Create account
+            </span>
+          </Link>
         </p>
-        {errors && <div className="text-red-600">{errors.root?.message}</div>}
-        <button
-          type="submit"
-          disabled={isSubmitting}
-          className="bg-green-primary-400 w-full rounded-lg py-3 text-white text-lg text-center cursor-pointer"
-        >
-          {isSubmitting ? 'Loading...' : 'Login'}
-        </button>
-      </form>
-      <p className="mt-10 mx-4">
-        Don’t have an account?
-        <Link to="/signup">
-          <span className="cursor-pointer text-secondary-blue mx-2 hover:text-blue-400">
-            Create account
-          </span>
-        </Link>
-      </p>
+      </div>
     </div>
   )
 }
