+++ src/pages/home/Home.tsx
@@ -1,0 +1,13 @@
+import Navbar from '../../components/Navbar'
+import HeroSection from './HeroSection'
+
+
+
+export default function Home() {
+  return (
+    <main>
+      <Navbar></Navbar>
+      <HeroSection />
+    </main>
+  )
+}
